+++ packages/contracts/contracts/votings/ERC20/ERC20Voting.sol
@@ -12,8 +12,9 @@ import "./../majority/MajorityVoting.sol";
 /// @notice The majority voting implementation using an ERC-20 token
 /// @dev This contract inherits from `MajorityVoting` and implements the `IMajorityVoting` interface
 contract ERC20Voting is MajorityVoting {
+    bytes4 internal constant ERC20_VOTING_INTERFACE_ID = MAJORITY_VOTING_INTERFACE_ID ^ this.getVotingToken.selector;
 
-    ERC20VotesUpgradeable public token;
+    ERC20VotesUpgradeable private token;
 
     /// @notice Initializes the component
     /// @dev This is required for the UUPS upgradability pattern
@@ -31,6 +32,7 @@ contract ERC20Voting is MajorityVoting {
         uint64 _minDuration,
         ERC20VotesUpgradeable _token
     ) public initializer {
+        _registerStandard(ERC20_VOTING_INTERFACE_ID);
         __MajorityVoting_init(
             _dao,
             _gsnForwarder,
@@ -40,6 +42,13 @@ contract ERC20Voting is MajorityVoting {
         );
 
         token = _token;
+    }
+
+    /// @notice getter function for the voting token
+    /// @dev public function also useful for registering interfaceId and for distinguishing from majority voting interface 
+    /// @return ERC20VotesUpgradeable the token used for voting 
+    function getVotingToken() public view returns(ERC20VotesUpgradeable){
+        return token;
     }
 
     /// @notice Returns the version of the GSN relay recipient