+++ packages/subgraph/src/utils/constants.ts
@@ -8,4 +8,5 @@ export const VOTER_STATE = new Map<number, string>()
   .set(3, 'Nay');
 
 export const MAJORITY_VOTING_INTERFACE = '0xc52cd5d9';
+export const ERC20_VOTING_INTERFACE = '0x27a0eec0';
 export const WHITELIST_VOTING_INTERFACE = '0x9dd60761';