+++ packages/contracts/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.0.1-alpha
 
 - First version of the package, exposing the JSON artifacts
+- implement getter function `getVotingToken` in `ERC20Voting.sol` and set `token` to private, to allow for registering ERC20Voting interfaceId.